+++ arctic_office_projects_api/__init__.py
@@ -1,8 +1,10 @@
 import logging
+from http import HTTPStatus
 
 import sentry_sdk
 
 from flask import Flask, request, has_request_context
+from flask import Flask, jsonify
 from flask.logging import default_handler
 
 from flask_reverse_proxy_fix.middleware import ReverseProxyPrefixFix
@@ -10,11 +12,11 @@ from flask_request_id_header.middleware import RequestID
 # noinspection PyPackageRequirements
 from werkzeug.exceptions import BadRequest, NotFound, InternalServerError, UnprocessableEntity
 
+from arctic_office_projects_api.utils import healthcheck_db
 from config import config
 from arctic_office_projects_api.extensions import db, auth
 from arctic_office_projects_api.meta.errors import error_handler_generic_bad_request, error_handler_generic_not_found, \
     error_handler_generic_internal_server_error, error_handler_generic_unprocessable_entity
-from arctic_office_projects_api.meta import meta as meta_blueprint
 from arctic_office_projects_api.main import main as main_blueprint
 
 
@@ -61,8 +63,52 @@ def create_app(config_name):
     app.register_error_handler(UnprocessableEntity, error_handler_generic_unprocessable_entity)
     app.register_error_handler(InternalServerError, error_handler_generic_internal_server_error)
 
-    # App
-    app.register_blueprint(meta_blueprint)
     app.register_blueprint(main_blueprint)
+    # Routes
+    app.add_url_rule('/', 'index', index_route)
+    app.add_url_rule('/meta/health/canary', 'canary_health_check', healthcheck_canary_route, methods=['get', 'options'])
+
+    # Blueprints
 
     return app
+
+
+# Route methods
+
+
+def index_route():
+    """
+    Returns a simple welcome message
+    """
+
+    payload = {
+        'meta': {
+            'summary': 'This API is used to record details of projects related to the NERC Arctic Office - '
+                       'https://www.arctic.ac.uk'
+        }
+    }
+
+    return jsonify(payload)
+
+
+def healthcheck_canary_route():
+    """
+    Returns whether this service is healthy
+
+    This healthcheck checks the application itself (assumed to be healthy if this method can be executed) and the
+    availability of its dependencies, such as databases.
+
+    If healthy a 204 No Content response is returned, if unhealthy a 503 Service Unavailable response is returned. This
+    healthcheck is binary and does not return any details to reduce payload size and prevent leaking sensitive data.
+
+    Other healthcheck's should be used where more details are required. This healthcheck is intended for use with load
+    balancers to give early indication of a service not being available.
+    """
+    dependencies = {
+        'db': healthcheck_db()
+    }
+
+    if False in dependencies.values():
+        return '', HTTPStatus.SERVICE_UNAVAILABLE
+
+    return '', HTTPStatus.NO_CONTENT